+++ indicators.py
@@ -81,7 +81,7 @@ def text_search(search_term, trommel_output):
 		cve_hit = '"(CVE-\d+-\d+ : .*\.)"'
 		name_hit = re.findall(cve_hit, search_text)
 		for match_hit in name_hit:
-			trommel_output.write("Check file version on embedded device [may need to emulate environment] - Found %s and it has been associated with %s\n" % (search_term, match_hit))
+			trommel_output.write("Found %s and it has been associated with %s\n" % (search_term, match_hit))
 	#Searches above CVE in Exploit-DB and Metasploit
 	for cve_hit in cve_field:
 		edb = exploitdb_result(cve_hit)
@@ -142,7 +142,7 @@ def kw(ff, trommel_output, names, bin_search):
 			drop_term = 'Dropbear server v[0-9]{4}\.[0-9]{2,3}'
 			drop_hit = re.search(drop_term, text)
 			if drop_hit:
-				trommel_output.write("The Dropbear (late 2011 or newer) binary found is %s\n" % drop_hit.group())
+				trommel_output.write("The Dropbear (late 2011 or newer) binary found is %s [may need to emulate environment]\n" % drop_hit.group())
 		text_search(dropbear_bin, trommel_output)
 	if telnet_bin in ff:
 		trommel_output.write("Non-Plain Text File, telnet binary file: %s\n" % ff)
@@ -153,10 +153,10 @@ def kw(ff, trommel_output, names, bin_search):
 	if busybox_bin in ff:
 		with open (ff, 'r') as keyword_search:
 			text = keyword_search.read()
-			bb_term = 'BusyBox v[0-9]{1}\.[0-9]{1,2}\.[0-9]{1}'
+			bb_term = 'BusyBox v[0-9]{1}\.[0-9]{1,2}\.[0-9]{1}|BusyBox v[0-9]{1}\.[0-9]{1,2}'
 			bb_hit = re.search(bb_term, text)
 			if bb_hit:
-				trommel_output.write("The BusyBox binary found is %s\n" % bb_hit.group())
+				trommel_output.write("The BusyBox binary found is %s [may need to emulate environment]\n" % bb_hit.group())
 		text_search(busybox_bin, trommel_output)
 	if other_bins in ff:
 		trommel_output.write("Non-Plain Text File, .bin file: %s\n" % ff)